Changed types of the database + Fixed showing of past activity + Fix #105
