+++ src/types/firestore/article.d.ts
@@ -1,0 +1,8 @@
+export interface Article {
+  __deleted: boolean
+  createdBy: string
+  creatorID: string
+  img: string
+  text: string
+  title: string
+}
