+++ src/types/firestore/category.d.ts
@@ -1,0 +1,9 @@
+export interface Category {
+  __deleted: boolean
+  id: string
+  bio: string
+  color: string
+  icon: string
+  name: string
+  filterName: string
+}
