+++ src/types/firestore/meal.d.ts
@@ -1,0 +1,13 @@
+export interface Meal {
+  createdBy: string
+  creatorID: string
+  id: string
+  img: string
+  ingredients: string
+  isActive: boolean
+  isHallal: boolean
+  isVegan: boolean
+  isVegetarian: boolean
+  name: string
+  price: string
+}
