+++ src/types/firestore/rule.d.ts
@@ -1,0 +1,8 @@
+export interface Rule {
+  __deleted: boolean
+  id: string
+  createdBy: string
+  creatorID: string
+  name: string
+  rule: string
+}
